brainag/requirements.txt
Django>=4.2.1,<5.1
psycopg2-binary>=2.9.6,<2.10
requests
pandas
plotly
djangorestframework
djangorestframework-simplejwt